move FormattedAppHeaderItems type out of global header directory
