+++ packages/gamut-labs/src/experimental/AppHeader/types.tsx
@@ -1,0 +1,12 @@
+import {
+  AppHeaderItem,
+} from './AppHeaderElements/types';
+
+export type FormattedAppHeaderItems = {
+  left: AppHeaderItem[];
+  right: AppHeaderItem[];
+};
+
+export type FormattedMobileAppHeaderItems = FormattedAppHeaderItems & {
+  mainMenu: AppHeaderItem[];
+};
